requirements.txt
Django>=4.2.0,<5.2.0
djangorestframework>=3.14.0
numpy>=1.23.0
scipy>=1.8.0
scikit-learn>=1.0.0
nltk>=3.7.0
networkx>=2.8.0
requests>=2.27.0
tqdm>=4.50.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0